# internal/http_server/handler/update_user_sub.go
package handler

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"task_manager/internal/http_server/dto"
	"task_manager/internal/lib/api/er"
	"task_manager/internal/lib/api/resp"
	valid "task_manager/internal/lib/api/valid"
	"task_manager/internal/lib/logger/sl"

	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-playground/validator/v10"
)

// UpdateSubscriptionHandler godoc
// @Summary      Update user subscription
// @Description  Updates a user's subscription data by its ID
// @Tags Subscription
// @Accept       json
// @Produce      json
// @Param        id   path      int  true  "Subscription ID"
// @Param        request body      dto.UpdateUserSubDTO true  "Data for updating the subscription"
// @Success      201  {object}  domain.UserSubscription
// @Failure      400  {object}  resp.ErrorResponse "Invalid ID or request body"
// @Failure      500  {object}  resp.ErrorResponse "Error updating subscription"
// @Router       /subscriptions/{id} [put]
func (h *UserSubscriptionHandler) UpdateSubscriptionHandler(w http.ResponseWriter, r *http.Request) {
	const op = "handler.AddUserSubscriptionHandler"

	ctx, cancel := context.WithTimeout(r.Context(), h.timeOut)
	defer cancel()

	log := h.log.With(
		slog.String("op", op),
		slog.String("request_url", middleware.GetReqID(r.Context())),
	)

	var req dto.UpdateUserSubDTO

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		log.Error("failed to decode request", sl.Err(err))

		resp.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}

	err := valid.ValidateDates(req.StartDate, req.EndDate)
	if err != nil {
		log.Error("invalid request body", sl.Err(err))

		resp.Error(w, fmt.Sprintf("invalid request body: %s", err), http.StatusBadRequest)
		return
	}

	validWithOpts := validator.New(validator.WithRequiredStructEnabled())
	if err := validWithOpts.Struct(req); err != nil {
		var validateErr validator.ValidationErrors
		errors.As(err, &validateErr)
		log.Error("invalid request", sl.Err(err))

		resp.Error(w, fmt.Sprintf("invalid request: %s", valid.ValidationError(validateErr, req)), http.StatusBadRequest)
		return
	}

	sub, err := h.service.UpdateById(ctx, req)
	if err != nil {
		log.Error("failed to update user subscription", sl.Err(err))
		if msg, code, ok := er.MapErrorToStatus(err); ok {
			resp.Error(w, msg, code)
			return
		}
		resp.Error(w, "failed to update user subscription", http.StatusInternalServerError)
		return
	}

	resp.ResponseOk(w, sub, http.StatusCreated)
}